Hold vs Move period
This is the implementation of analysis to compare the neural activity separation for hold and move sections of CST. The neural data showed clear separation for control strategy, but it was strongly correlated with trial order as well. To rule out that the separation we observe is due to drift of activity over trials, we looked at hold period where no movement is done, and tested whether there is still neural separation between early and late trials, which doesn't seem to be the case.